shutdown: add -r for reboot

+++ skeleton/.tmp_update/runtime.sh
@@ -14,6 +14,5 @@ if [ -f "$LAUNCH_PATH" ]; then
 fi
 
 while true; do
-	reboot
-	sleep 10
+	sync && reboot && sleep 10
 done
